translation seems to work, but issue with weights
translation like with rotation, weight vector is size 1 ( changed to 21 but not sure correct windows yet).  The problem in translation case, is the weight normalization is still performed so results weight == 1.0 .  Maybe that is correct but need to verify.  Still for test data (from phaser repo) the source_3, target_3 produced translation of ~0.5 meters.  Which appears correct in blender.